Conflicts:
	simhalt.cc
	simversion.h

+++ simfab.cc
@@ -2886,8 +2886,8 @@ void fabrik_t::calc_max_intransit_percentages()
 			continue;
 		}
 		const uint16 time_to_consume = max(1, get_time_to_consume_stock(index)); 
-		const uint32 ratio = (lead_time * 1000 / time_to_consume);
-		const uint32 modified_max_intransit_percentage = (ratio * base_max_intransit_percentage) / 1000;
+		const uint32 ratio = ((uint32)lead_time * 1000 / (uint32)time_to_consume);
+		const uint32 modified_max_intransit_percentage = (ratio * (uint32)base_max_intransit_percentage) / 1000;
 		max_intransit_percentages.put(catg, (uint16)modified_max_intransit_percentage);
 		index ++;
 	}
